+++ envoy-control-core/src/test/kotlin/pl/allegro/tech/servicemesh/envoycontrol/groups/MetadataNodeGroupTest.kt
@@ -3,15 +3,22 @@ package pl.allegro.tech.servicemesh.envoycontrol.groups
 import com.google.protobuf.Struct
 import com.google.protobuf.Value
 import com.google.protobuf.util.Durations
+import com.google.protobuf.util.Values
 import io.envoyproxy.envoy.config.accesslog.v3.ComparisonFilter
+import io.envoyproxy.envoy.config.core.v3.BuildVersion
+import io.envoyproxy.envoy.config.core.v3.Node
+import io.envoyproxy.envoy.type.v3.SemanticVersion
 import io.grpc.Status
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Assertions.assertFalse
+import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import pl.allegro.tech.servicemesh.envoycontrol.groups.CommunicationMode.ADS
 import pl.allegro.tech.servicemesh.envoycontrol.groups.CommunicationMode.XDS
+import pl.allegro.tech.servicemesh.envoycontrol.groups.ListenersConfig.AddUpstreamServiceTagsCondition
 import pl.allegro.tech.servicemesh.envoycontrol.snapshot.SnapshotProperties
 import pl.allegro.tech.servicemesh.envoycontrol.snapshot.serviceDependencies
 import io.envoyproxy.envoy.config.core.v3.Node as NodeV3
@@ -257,7 +264,10 @@ class MetadataNodeGroupTest {
         val nodeGroup = MetadataNodeGroup(createSnapshotProperties())
         val metadata = createMetadataBuilderWithDefaults()
 
-        metadata!!.putFields("access_log_filter", accessLogFilterProto(value = "EQ:400", fieldName = "status_code_filter"))
+        metadata!!.putFields(
+            "access_log_filter",
+            accessLogFilterProto(value = "EQ:400", fieldName = "status_code_filter")
+        )
 
         // when
         val group = nodeGroup.hash(NodeV3.newBuilder().setMetadata(metadata.build()).build())
@@ -279,7 +289,10 @@ class MetadataNodeGroupTest {
         val nodeGroup = MetadataNodeGroup(createSnapshotProperties())
         val metadata = createMetadataBuilderWithDefaults()
 
-        metadata!!.putFields("access_log_filter", accessLogFilterProto(value = config, fieldName = "status_code_filter"))
+        metadata!!.putFields(
+            "access_log_filter",
+            accessLogFilterProto(value = config, fieldName = "status_code_filter")
+        )
 
         // expect
         val exception = assertThrows<NodeMetadataValidationException> {
@@ -303,6 +316,54 @@ class MetadataNodeGroupTest {
         assertThat(group.listenersConfig!!.accessLogFilterSettings.statusCodeFilterSettings).isNull()
     }
 
+    @Test
+    fun `should correctly compare envoy version`() {
+        // given
+        val v1dot21 = envoyVersion(1, 21)
+        val v1dot24 = envoyVersion(1, 24)
+        val v1dot26 = envoyVersion(1, 26)
+        val v2dot21 = envoyVersion(2, 21)
+
+        // expect
+        assertTrue(v1dot21 < v1dot24)
+        assertFalse(v1dot24 < v1dot24)
+        assertFalse(v1dot26 < v1dot24)
+        assertFalse(v2dot21 < v1dot24)
+    }
+
+    @Test
+    fun `should set addUpstreamServiceTag condition according to client flag and envoy version`() {
+        // given
+        val nodeGroup = MetadataNodeGroup(createSnapshotProperties())
+
+        fun assertThat(node: Node, has: AddUpstreamServiceTagsCondition) {
+            // when
+            val group = nodeGroup.hash(node)
+
+            // then
+            assertThat(group.listenersConfig.orDefault().addUpstreamServiceTags).isEqualTo(has)
+        }
+
+        val emptyNode = NodeV3.newBuilder().build()
+        val basicNode = NodeV3.newBuilder().setMetadata(createMetadataBuilderWithDefaults()).build()
+        val newEnvoyNode = basicNode.with(envoyVersion(1, 24))
+        val oldEnvoyNode = basicNode.with(envoyVersion(1, 23))
+        val enabledNewEnvoyNode = newEnvoyNode.withMetadata("add_upstream_service_tags", Values.of(true))
+        val enabledOldEnvoyNode = oldEnvoyNode.withMetadata("add_upstream_service_tags", Values.of(true))
+        val disabledNewEnvoyNode = newEnvoyNode.withMetadata("add_upstream_service_tags", Values.of(false))
+        val disabledOldEnvoyNode = oldEnvoyNode.withMetadata("add_upstream_service_tags", Values.of(false))
+
+        // expect
+        assertThat(node = emptyNode, has = AddUpstreamServiceTagsCondition.NEVER)
+        assertThat(node = basicNode, has = AddUpstreamServiceTagsCondition.NEVER)
+        assertThat(node = newEnvoyNode, has = AddUpstreamServiceTagsCondition.WHEN_SERVICE_TAG_PREFERENCE_IS_USED)
+        assertThat(node = oldEnvoyNode, has = AddUpstreamServiceTagsCondition.NEVER)
+        assertThat(node = enabledNewEnvoyNode, has = AddUpstreamServiceTagsCondition.ALWAYS)
+        assertThat(node = enabledOldEnvoyNode, has = AddUpstreamServiceTagsCondition.NEVER)
+        assertThat(node = disabledNewEnvoyNode, has = AddUpstreamServiceTagsCondition.NEVER)
+        assertThat(node = disabledOldEnvoyNode, has = AddUpstreamServiceTagsCondition.NEVER)
+    }
+
     private fun createMetadataBuilderWithDefaults(): Struct.Builder? {
         val metadata = NodeV3.newBuilder().metadataBuilder
         metadata.putFields("ingress_host", Value.newBuilder().setStringValue("127.0.0.1").build())
@@ -323,4 +384,12 @@ class MetadataNodeGroupTest {
         snapshotProperties.incomingPermissions.enabled = incomingPermissions
         return snapshotProperties
     }
+
+    private fun NodeV3.with(version: SemanticVersion) = toBuilder()
+        .setUserAgentBuildVersion(BuildVersion.newBuilder().setVersion(version))
+        .build()
+
+    private fun NodeV3.withMetadata(key: String, value: Value) = toBuilder()
+        .apply { metadataBuilder.putFields(key, value) }
+        .build()
 }